# schemas/api/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactType": {
        "description": "Mirrors ``schemas/ontology.yaml: artifact_types``. ``SUMMARY`` is produced by the engine, not\nby the LLM artifact list, so it is absent from ``episode_extraction.schema.json``.",
        "enum": [
          "decision",
          "requirement",
          "task",
          "finding",
          "hypothesis",
          "experiment",
          "summary"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
      "ArtifactView": {
        "additionalProperties": false,
        "description": "A ``knowledge_artifacts`` row with its supersession chain (decisions, tasks, findings, ...).",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "citation": {
            "default": "",
            "title": "Citation",
            "type": "string"
          },
          "confidence": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "entity_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "evidence_quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Quote"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_status": {
            "default": "active",
            "title": "Source Status",
            "type": "string"
          },
          "status": {
            "default": "current",
            "title": "Status",
            "type": "string"
          },
          "structured": {
            "additionalProperties": true,
            "title": "Structured",
            "type": "object"
          },
          "superseded_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By Id"
          },
          "superseded_by_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By Title"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Id"
          },
          "supersedes_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Title"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "track": {
            "default": "unknown",
            "title": "Track",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body"
        ],
        "title": "ArtifactView",
        "type": "object"
      },
...
      "DecisionCreate": {
        "additionalProperties": false,
        "description": "``POST /v1/decisions`` body. ``supersedes_id`` adds the ADR-0005 rule 6 link, never a delete.",
        "properties": {
          "body": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Id"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body"
        ],
        "title": "DecisionCreate",
        "type": "object"
      },
...
      "EpisodeCreate": {
        "additionalProperties": false,
        "description": "``POST /v1/episodes`` body. ``client`` tags the writer (ADR-0008 audit requirement).",
        "properties": {
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MCP client id; omitted for a manual write",
            "title": "Client"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "text"
        ],
        "title": "EpisodeCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObjectType": {
        "description": "Polymorphic target of ``embeddings``, ``retrieval_logs``, ``extraction_reviews`` and\n``Gateway.explain(id)``.",
        "enum": [
          "chunk",
          "artifact",
          "fact",
          "entity",
          "episode",
          "source"
        ],
        "title": "ObjectType",
        "type": "string"
      },
...
      "Provenance": {
        "additionalProperties": false,
        "description": "The provenance stamp carried by every derived row.\n\n``source_version`` is the ``source_versions.id`` the row was derived from - the plan's column\nname is kept verbatim even though it holds a version *id*, because A04's DDL, the\n``provenance_v`` view and the MCP payload all use that name.\n\nThe three chunk-only fields (``heading_path``, ``char_start``, ``char_end``) are optional and set\nonly when the provenance points at a chunk.",
        "properties": {
          "char_end": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "confidence": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "device_id": {
            "description": "devices.id; 'local-development-machine' in V0.1",
            "title": "Device Id",
            "type": "string"
          },
          "embedding_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "embedding_models.id",
            "title": "Embedding Model Id"
          },
          "episode_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "extraction_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "extraction_models.id, or 'deterministic:registry-v1'",
            "title": "Extraction Model Id"
          },
          "heading_path": {
            "items": {
              "type": "string"
            },
            "title": "Heading Path",
            "type": "array"
          },
          "ingestion_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingestion Run Id"
          },
          "observed_at": {
            "description": "Source time (mtime/commit) or run time (ADR-0005 #6)",
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "projects.id slug",
            "title": "Project Id"
          },
          "source_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "sha256: content hash of the version",
            "title": "Source Hash"
          },
          "source_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "sources.id",
            "title": "Source Id"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Logical URI (ADR-0004)",
            "title": "Source Uri"
          },
          "source_version": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "source_versions.id",
            "title": "Source Version"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "NULL = still current",
            "title": "Valid To"
          }
        },
        "required": [
          "device_id",
          "observed_at"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WriteReceipt": {
        "additionalProperties": false,
        "description": "ADR-0008 append-only write result. ``accepted=False`` carries the reason, never a stack.",
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "episode_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "object_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "object_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObjectType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "accepted"
        ],
        "title": "WriteReceipt",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local personal memory infrastructure (V0.1). Hybrid retrieval over a personal corpus with provenance on every returned object, point-in-time reads, and append-only writes that are disabled by default (ADR-0008). Published on 127.0.0.1 only (ADR-0007).",
    "title": "AI Memory Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/decisions": {
      "get": {
        "operationId": "get_decisions_v1_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "include_superseded",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Superseded",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactView"
                  },
                  "title": "Response Get Decisions V1 Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision artifacts",
        "tags": [
          "read"
        ]
      },
      "post": {
        "operationId": "record_decision_v1_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append a decision artifact (403 unless GATEWAY_WRITE_ENABLED)",
        "tags": [
          "write"
        ]
      }
    },
...
    "/v1/episodes": {
      "post": {
        "operationId": "add_episode_v1_episodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append an episode (403 unless GATEWAY_WRITE_ENABLED)",
        "tags": [
          "write"
        ]
      }
    },
...
  }
}